Adding the api modules to generate v1 endpoints
