+++ eox_core/api/urls.py
@@ -5,6 +5,5 @@ from eox_core import views
 
 
 urlpatterns = [  # pylint: disable=invalid-name
-    url(r'^eox-info$', views.info_view),
-    url(r'^api/', include('eox_core.api.urls', namespace='eox-api')),
+    url(r'^v1/', include('eox_core.api.v1.urls', namespace='eox-api')),
 ]
